+++ packages/adapters/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.23.3](https://github.com/dcmjs-org/dcmjs/compare/v1.23.2...v1.23.3) (2023-10-20)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.23.2](https://github.com/dcmjs-org/dcmjs/compare/v1.23.1...v1.23.2) (2023-10-20)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.23.1](https://github.com/dcmjs-org/dcmjs/compare/v1.23.0...v1.23.1) (2023-10-20)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [1.23.0](https://github.com/dcmjs-org/dcmjs/compare/v1.22.1...v1.23.0) (2023-10-19)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.22.1](https://github.com/dcmjs-org/dcmjs/compare/v1.22.0...v1.22.1) (2023-10-19)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [1.22.0](https://github.com/dcmjs-org/dcmjs/compare/v1.21.2...v1.22.0) (2023-10-17)
+
+### Features
+
+-   **colormap:** add colormap props and default properties ([#834](https://github.com/dcmjs-org/dcmjs/issues/834)) ([475914d](https://github.com/dcmjs-org/dcmjs/commit/475914d0eaa35f1ae65b989c74efda042dc6d97a))
+
+## [1.21.2](https://github.com/dcmjs-org/dcmjs/compare/v1.21.1...v1.21.2) (2023-10-16)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.21.1](https://github.com/dcmjs-org/dcmjs/compare/v1.21.0...v1.21.1) (2023-10-14)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [1.21.0](https://github.com/dcmjs-org/dcmjs/compare/v1.20.3...v1.21.0) (2023-10-10)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.20.3](https://github.com/dcmjs-org/dcmjs/compare/v1.20.2...v1.20.3) (2023-10-09)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.20.2](https://github.com/dcmjs-org/dcmjs/compare/v1.20.1...v1.20.2) (2023-10-09)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
 ## [1.20.1](https://github.com/dcmjs-org/dcmjs/compare/v1.20.0...v1.20.1) (2023-10-06)
 
 ### Bug Fixes